Auto detect and create paths for all direct-attached volumes with drive letters assigned

+++ PowerShell/Create-RestoreFile.ps1
@@ -6,12 +6,18 @@
               File contains just a date/time entry for when it was created.
 #>
 
-$folder = "C:\BackupTests\SourceFile"
-$file = "RestoreMe.txt"
+$drives = Get-WmiObject -Class Win32_Volume | ?{$_.DriveType -eq 3}
 
-If (!(Test-Path $folder)) {New-Item -ItemType "directory" -Path $folder}
-If (Test-Path (Join-Path -Path $folder -ChildPath $file)) {
-    Remove-Item -Path (Join-Path -Path $folder -ChildPath $file) -Force
+foreach ($drive in $drives) {
+    If ($drive.DriveLetter[0] -ne $null) {
+        $folder = $drive.DriveLetter + "\BackupTests\SourceFile"
+        $file = "RestoreMe.txt"
+
+        If (!(Test-Path $folder)) {New-Item -ItemType "directory" -Path $folder}
+        If (Test-Path (Join-Path -Path $folder -ChildPath $file)) {
+            Remove-Item -Path (Join-Path -Path $folder -ChildPath $file) -Force
+        }
+
+        New-Item -Path $folder -ItemType "file" -Name $file -Value (Get-Date)
+    }
 }
-
-New-Item -Path $folder -ItemType "file" -Name $file -Value (Get-Date)
